requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
black==25.1.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.13
flake8==7.3.0
graphql-core==3.2.6
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
isort==6.0.1
lia-web==0.2.2
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platform_common @ git+https://${PLATFORM_COMMON_TOKEN}@github.com/migoVanDingo/ed-platform-common@442525a9cc5c256eaaa34ccbb52db1ac518d0c8c
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.3.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
strawberry-graphql==0.278.1
structlog==25.4.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ulid-py==1.1.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
